refactor(evm): extract shared trySignEip2612Permit and trySignErc20Approval from scheme classes
