+++ typescript/packages/mechanisms/evm/src/shared/permit2-helpers.ts
@@ -1,5 +1,6 @@
 import {
   PaymentPayload,
+  PaymentPayloadContext,
   PaymentPayloadResult,
   PaymentRequirements,
   FacilitatorContext,
@@ -10,6 +11,7 @@ import {
   extractEip2612GasSponsoringInfo,
   validateEip2612GasSponsoringInfo,
   extractErc20ApprovalGasSponsoringInfo,
+  EIP2612_GAS_SPONSORING_KEY,
   ERC20_APPROVAL_GAS_SPONSORING_KEY,
   type Eip2612GasSponsoringInfo,
   type Erc20ApprovalGasSponsoringFacilitatorExtension,
@@ -30,6 +32,9 @@ import {
 import { ClientEvmSigner, FacilitatorEvmSigner } from "../signer";
 import { ExactPermit2Payload, Permit2Authorization, UptoPermit2Payload } from "../types";
 import { validateErc20ApprovalForPayment } from "../exact/facilitator/erc20approval";
+import { signEip2612Permit } from "../exact/client/eip2612";
+import { signErc20ApprovalTransaction } from "../exact/client/erc20approval";
+import { resolveExtensionRpcCapabilities, type ExactEvmSchemeOptions } from "../exact/client/rpc";
 
 /**
  * Base type for Permit2 payloads shared between exact and upto schemes.
@@ -536,4 +541,160 @@ async function signPermit2Authorization(
       },
     },
   });
+}
+
+// ---------------------------------------------------------------------------
+// Client-side extension helpers (shared by ExactEvmScheme and UptoEvmScheme)
+// ---------------------------------------------------------------------------
+
+/**
+ * Attempts to sign an EIP-2612 permit for gasless Permit2 approval.
+ *
+ * Returns extension data if:
+ * 1. Server advertises eip2612GasSponsoring
+ * 2. Signer has readContract capability
+ * 3. Current Permit2 allowance is insufficient
+ *
+ * @param signer - The EVM client signer
+ * @param options - Optional RPC configuration for backfilling capabilities
+ * @param requirements - The payment requirements from the server
+ * @param result - The payment payload result from the scheme
+ * @param context - Optional context containing server extensions and metadata
+ * @returns Extension data for EIP-2612 gas sponsoring, or undefined if not applicable
+ */
+export async function trySignEip2612PermitExtension(
+  signer: ClientEvmSigner,
+  options: ExactEvmSchemeOptions | undefined,
+  requirements: PaymentRequirements,
+  result: PaymentPayloadResult,
+  context?: PaymentPayloadContext,
+): Promise<Record<string, unknown> | undefined> {
+  const capabilities = resolveExtensionRpcCapabilities(requirements.network, signer, options);
+
+  if (!capabilities.readContract) {
+    return undefined;
+  }
+
+  if (!context?.extensions?.[EIP2612_GAS_SPONSORING_KEY]) {
+    return undefined;
+  }
+
+  const tokenName = requirements.extra?.name as string | undefined;
+  const tokenVersion = requirements.extra?.version as string | undefined;
+  if (!tokenName || !tokenVersion) {
+    return undefined;
+  }
+
+  const chainId = getEvmChainId(requirements.network);
+  const tokenAddress = getAddress(requirements.asset) as `0x${string}`;
+
+  try {
+    const allowance = (await capabilities.readContract({
+      address: tokenAddress,
+      abi: erc20AllowanceAbi,
+      functionName: "allowance",
+      args: [signer.address, PERMIT2_ADDRESS],
+    })) as bigint;
+
+    if (allowance >= BigInt(requirements.amount)) {
+      return undefined;
+    }
+  } catch {
+    // Allowance check failed, proceed with signing
+  }
+
+  const permit2Auth = result.payload?.permit2Authorization as Record<string, unknown> | undefined;
+  const deadline =
+    (permit2Auth?.deadline as string) ??
+    Math.floor(Date.now() / 1000 + requirements.maxTimeoutSeconds).toString();
+
+  const info = await signEip2612Permit(
+    {
+      address: signer.address,
+      signTypedData: msg => signer.signTypedData(msg),
+      readContract: capabilities.readContract,
+    },
+    tokenAddress,
+    tokenName,
+    tokenVersion,
+    chainId,
+    deadline,
+    requirements.amount,
+  );
+
+  return {
+    [EIP2612_GAS_SPONSORING_KEY]: { info },
+  };
+}
+
+/**
+ * Attempts to sign an ERC-20 approval transaction for gasless Permit2 approval.
+ *
+ * This is the fallback path when the token does not support EIP-2612. The client
+ * signs (but does not broadcast) a raw `approve(Permit2, MaxUint256)` transaction.
+ * The facilitator broadcasts it atomically before settling.
+ *
+ * Returns extension data if:
+ * 1. Server advertises erc20ApprovalGasSponsoring
+ * 2. Signer has signTransaction + getTransactionCount capabilities
+ * 3. Current Permit2 allowance is insufficient
+ *
+ * @param signer - The EVM client signer
+ * @param options - Optional RPC configuration for backfilling capabilities
+ * @param requirements - The payment requirements from the server
+ * @param context - Optional context containing server extensions and metadata
+ * @returns Extension data for ERC-20 approval gas sponsoring, or undefined if not applicable
+ */
+export async function trySignErc20ApprovalExtension(
+  signer: ClientEvmSigner,
+  options: ExactEvmSchemeOptions | undefined,
+  requirements: PaymentRequirements,
+  context?: PaymentPayloadContext,
+): Promise<Record<string, unknown> | undefined> {
+  const capabilities = resolveExtensionRpcCapabilities(requirements.network, signer, options);
+
+  if (!capabilities.readContract) {
+    return undefined;
+  }
+
+  if (!context?.extensions?.[ERC20_APPROVAL_GAS_SPONSORING_KEY]) {
+    return undefined;
+  }
+
+  if (!capabilities.signTransaction || !capabilities.getTransactionCount) {
+    return undefined;
+  }
+
+  const chainId = getEvmChainId(requirements.network);
+  const tokenAddress = getAddress(requirements.asset) as `0x${string}`;
+
+  try {
+    const allowance = (await capabilities.readContract({
+      address: tokenAddress,
+      abi: erc20AllowanceAbi,
+      functionName: "allowance",
+      args: [signer.address, PERMIT2_ADDRESS],
+    })) as bigint;
+
+    if (allowance >= BigInt(requirements.amount)) {
+      return undefined;
+    }
+  } catch {
+    // Allowance check failed, proceed with signing
+  }
+
+  const info = await signErc20ApprovalTransaction(
+    {
+      address: signer.address,
+      signTransaction: capabilities.signTransaction,
+      getTransactionCount: capabilities.getTransactionCount,
+      estimateFeesPerGas: capabilities.estimateFeesPerGas,
+    },
+    tokenAddress,
+    chainId,
+  );
+
+  return {
+    [ERC20_APPROVAL_GAS_SPONSORING_KEY]: { info },
+  };
 }
